Use DebugAdapterTracker to get debug pid
It now is able to get the correct PID for Python by using the DebugAdapterTracker to read messages between VSCode and the Python Debug Adapter.

Unfortunately, this does not work for other debuggers such as NodeJS, so a separate universal solution must be found, or each language must be supported individually.